requirements.txt
requests
uvicorn
pydantic
pydantic_core
pydantic_settings
passlib
python-jose
bcrypt
fastapi
python-dotenv
python-multipart
prometheus_client
prometheus_fastapi_instrumentator
pytest-mock
pytest-cov
openai
langchain_xai
chainlit